Added ones and zeros pretraining
@Rafipilot note how it's similar to how we pretrained the clam/pavolv agents. Let's see what this does.

Another option is to move away from Cpos and Cneg, using instead next_state with all zeros or all ones as a label. Or maybe all 4 options-- so Cpos and Cneg can be when the user likes/dislikes, and all ones / all zeros when the user is very certain they like or dislike. We'll experiment and discuss.

Co-Authored-By: Rafayel Latif <74052861+Rafipilot@users.noreply.github.com>